Firedrake backend: Move parameters code
