+++ tlm_adjoint/firedrake/__init__.py
@@ -16,6 +16,11 @@ from .equations import *            # noqa: E402,F401
 from .functions import *            # noqa: E402,F401
 from .hessian_system import *       # noqa: E402,F401
 from .interpolation import *        # noqa: E402,F401
+from .parameters import *           # noqa: E402,F401
 from .projection import *           # noqa: E402,F401
 
 from .backend_patches import *      # noqa: E402,F401
+
+del backend, backend_code_generator_interface, backend_interface, assignment, \
+    block_system, caches, equations, functions, hessian_system, \
+    interpolation, parameters, projection, backend_patches  # noqa: F821
